refactored(SPV-813): per review comments client methods been moved to pike.go and used post
